Fixing mistake of deleting __init__.py

+++ apps/backend/rolemanagment/views.py
@@ -1,14 +1,11 @@
-from django.contrib.auth import login, logout
+from database.models import User, Teacher
 from rest_framework import status, permissions
 from rest_framework import viewsets
-from rest_framework.authentication import SessionAuthentication, TokenAuthentication
 from rest_framework.response import Response
 from rest_framework.views import APIView
-from rest_framework.decorators import api_view
-from database.models import Teacher, User
-from .serializers import UserSerializer, EditRoleSerializer
-from rest_framework.generics import DestroyAPIView
+
 from .permissions import IsSuperUser
+from .serializers import UserSerializer, TeacherSerializer
 
 
 class UserView(viewsets.ModelViewSet, APIView):
@@ -33,25 +30,18 @@ class UserView(viewsets.ModelViewSet, APIView):
         return Response(serializer.data, status=status.HTTP_200_OK)
 
 
-class EditRoleView(viewsets.ModelViewSet, APIView):
-
+class TeacherView(viewsets.ModelViewSet, APIView):
+    """
+    API endpoint that allows users to be viewed or edited.
+    """
     permission_classes = (permissions.AllowAny, IsSuperUser,)
 
-    serializer_class = EditRoleSerializer
+    serializer_class = TeacherSerializer
 
     queryset = Teacher.objects.all()
 
-    def get(self, request):
-        data = Teacher.objects.all()
-        serializer = EditRoleSerializer(data, many=True)
-        return Response(serializer.data)
-
-    def post(self, request):
-        serializer = EditRoleSerializer(data=request.data)
-        if serializer.is_valid():
-            serializer.save()
-            return Response(serializer.data, status=status.HTTP_201_CREATED)
-        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+    def get_queryset(self):
+        return Teacher.objects.all()
 
     def patch(self, request, pk):
         instance = self.get_object(pk=pk)
@@ -61,10 +51,47 @@ class EditRoleView(viewsets.ModelViewSet, APIView):
         serializer.save()
         return Response(serializer.data, status=status.HTTP_200_OK)
 
-    def delete(self, request, pk):
-        try:
-            teacher = Teacher.objects.get(pk=pk)
-            teacher.delete()
-            return Response(status=status.HTTP_204_NO_CONTENT)
-        except Teacher.DoesNotExist:
-            return Response(status=status.HTTP_404_NOT_FOUND)
+
+class EditRoleView(viewsets.ModelViewSet, APIView):
+    """
+    API endpoint that allows users to be viewed or edited.
+    """
+    permission_classes = (permissions.AllowAny, IsSuperUser,)
+
+    # Patch that change the role of a user based on the id and the role received in the request
+    # If the role is 'professor' or 'tutor' the user is added to the teacher table as a professor (professor = True and tutor = False) or as a tutor (professor = False and tutor = True) and the is_staff field of the user is set to True
+    # If the role is 'student' and the user is in the teacher table, the user is removed from the teacher table and the superuser field is set to False
+    # If the role is 'superuser' the is_superuser field of the user is set to True and the user is added to the teacher table as a professor
+    def patch(self, request):
+        user_id = request.data.get('id')
+        role = request.data.get('role')
+        user = User.objects.get(id=user_id)
+        if role == 'professor' or role == 'tutor':
+            user.is_staff = True
+            user.save()
+            if Teacher.objects.filter(user=user).exists():
+                teacher = Teacher.objects.get(user=user)
+                teacher.is_professor = True if role == 'professor' else False
+                teacher.is_tutor = True if role == 'tutor' else False
+                teacher.save()
+            else:
+                Teacher.objects.create(user=user, is_professor=True if role == 'professor' else False,
+                                       is_tutor=True if role == 'tutor' else False)
+        elif role == 'student':
+            if Teacher.objects.filter(user=user).exists():
+                teacher = Teacher.objects.get(user=user)
+                teacher.delete()
+                user.is_superuser = False
+                user.save()
+        elif role == 'superuser':
+            user.is_superuser = True
+            user.save()
+            if Teacher.objects.filter(user=user).exists():
+                teacher = Teacher.objects.get(user=user)
+                teacher.is_professor = True
+                teacher.is_tutor = False
+                teacher.save()
+            else:
+                Teacher.objects.create(user=user, is_professor=True, is_tutor=False)
+
+        return Response({}, status=status.HTTP_201_CREATED)
